pyproject.toml
[tool.poetry]
name = "pharmapi"
version = "0.1.0"
description = ""
authors = ["zubairmh <zubair.mh@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
motor = "^3.4.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
